+++ packages/vda-sbt-client/CHANGELOG.md
@@ -1,8 +1,17 @@
+2023-03-13 (v2.1.5)
+-------------------
+- Updated `isSBTClaimed()` function
+- Added `getClaimedSBTList()` function
+- Updated test code
+
+2023-03-07 (v2.1.4)
+-------------------
 
 - Improve documentation in README.md
 - Simplify usage
 - Improve code comments
 - Add error reason for failed claims
+- Fix bug with signing mint request
 
 2023-02-09 (v2.0.5)
 -------------------